Nightwork - Finally fixed annoying thing <Description>
Since I added some images the app kept closing activities to clear the
memory. This is quite understandable and likely to happen with all the
apps we make.
There are a couple of ways to fix this problem. At first I tried to use
savedInstanceState which seems the most logical and legit choice. This
wasn't happening because of a lack of expertise and experience with
implementing the Parcable interface and structure throughout the
project. After having several shots at it and numerous rollbacks I tried
something else.

I warn you, this ain't pretty.

The problem was that the core activity was being destroyed and when it
was created again > there was a complete new game initialized. AKA all
relevant data was gone!

My solution (again, ain't pretty)
Override the CoreActivity finish() method!
The finish method does nothing now, except when the back button is
clicked. So we have to find a better solution, but for the demo this
seemed sufficient to me and there are comments near everything
discutable.

Anyway, time to go to bed soon. Just checking some stuff and maybe
tweaking a couple of things.

Peace!